+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF0V8C5WHGFNKNNB9FFZVNCR"
+        "01KF1769YPG7974TB7W27T7X6H"
     )
     r"""The unique identifier of the data source"""