+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kf0v8bphnq7bce166b3ndhez"
+        "conv_01kf1769gx4avd36tcpt26mbhz"
     )
     r"""Unique conversation identifier with `conv_` prefix."""