+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -194,7 +194,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-15T12:49:16.473Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-15T16:17:52.624Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")